Move webhook test in an individual file
